requirements.txt
Flask>=2.0
requests>=2.20
python-dotenv>=0.15
vercel-kv==0.1.5
